+++ Controls/_filterPanel.less
@@ -1,4 +1,4 @@
 @import "Controls/_filterPanel/View/_View";
 @import "Controls/_filterPanel/Editors/_List";
 @import "Controls/_filterPanel/Editors/_NumberRange";
-@import "Controls/_filterPanel/Editors/_ToggleTumbler";
+@import "Controls/_filterPanel/Editors/_Tumbler";
