+++ Controls/_filterPanel/Editors/Tumbler.wml
@@ -1,7 +1,7 @@
 <div class="controls-FilterEditors__toggleTumbler_theme-{{_options.theme}}">
    <Controls.toggle:Tumbler items="{{_options.items}}"
+                            selectedKey="{{_options.propertyValue}}"
                             readOnly="{{_options.readOnly}}"
-                            bind:selectedKey="_selectedKey"
                             on:selectedKeyChanged="_selectedKeyChangedHandler()"/>
 </div>
