+++ Controls/_filterPanel/Editors/_Tumbler.less
